requirements.txt
gunicorn==19.6.0
alembic==0.8.6
blinker==1.4
click==6.6
decorator==4.0.9
et-xmlfile==1.0.1
Flask==0.11.1
Flask-Admin==1.4.0
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
itsdangerous==0.24
jdcal==1.2
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
pbr==1.9.1
PyMySQL==0.7.2
python-editor==1.0
selenium==2.53.2
six==1.10.0
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
Werkzeug==0.11.10
Whoosh==2.7.4
WTForms==2.1
